breaking the internal classes of human formatter into different files.
