+++ lib/active_model/better_errors/formatter/human/translator/attribute.rb
@@ -1,0 +1,15 @@
+module ActiveModel
+  module BetterErrors
+    class Formatter
+      class Human
+        class Translator
+          class Attribute < self
+            def translate
+              Builder::Attribute.build(error_message)
+            end
+          end
+        end
+      end
+    end
+  end
+end
